rendering of claims screen refactored
